Service - InstructorLectureJpaService getUpdateREmainingPersonnel parameter type 변경으로 인한 typeCasting 수정

+++ src/main/java/basic/classroom/service/members/instructor/InstructorLectureJpaService.java
@@ -159,7 +159,7 @@ public class InstructorLectureJpaService {
         }
     }
     private int getUpdateRemainingPersonnel(Lecture lecture, UpdateLectureRequest updateLectureRequest) {
-        int updatePersonnel = updateLectureRequest.getPersonnel();
+        int updatePersonnel = updateLectureRequest.getPersonnel().intValue();
         int remainingPersonnel = lecture.getRemainingPersonnel();
         int personnel = lecture.getPersonnel();
         int appliedPersonnel = personnel - remainingPersonnel;
